fix(types): remove unnecessary structure of folder and files
