+++ src/components/data/Icon/interfaces/Icon.interface.ts
@@ -1,8 +1,7 @@
 import { Appearance } from "@src/shared/types/Appearance.type";
-import { Shape } from "../types/Icon.Shape.type";
-import { Spacing } from "../types/Icon.Spacing.type";
-import { Variant } from "../types/Icon.Variant.type";
+
 import { inube } from "@src/shared/tokens";
+import { Shape, Spacing, Variant } from "../stories/props";
 
 export interface IIconProps {
   appearance: Appearance;
