--- src/components/data/Icon/types/Icon.Shape.type.ts
@@ -1,2 +1,0 @@
-export const shapes = ["circle", "rectangle"] as const;
-export type Shape = typeof shapes[number];
