--- src/components/data/Icon/types/Icon.Spacing.type.ts
@@ -1,2 +1,0 @@
-export const spacings = ["none", "compact", "wide"] as const;
-export type Spacing = typeof spacings[number];
