--- src/components/data/Icon/types/Icon.Variant.type.ts
@@ -1,2 +1,0 @@
-export const variants = ["filled", "outlined", "none"] as const;
-export type Variant = typeof variants[number];
